✨ model(trace): Add TraceSpan and TraceResponse types
Move trace models from routes/traces.rs to models/trace.rs with proper
DateTime<Utc> timestamps, Deserialize support, and parent_span_id field.
Update OpenAPI schema references to use the new model location.
