+++ aa-api/src/models/mod.rs
@@ -2,6 +2,8 @@
 
 pub mod event;
 pub mod event_type;
+pub mod trace;
 
 pub use event::{EventId, GovernanceEvent};
 pub use event_type::EventType;
+pub use trace::{TraceResponse, TraceSpan};
